# src/main/java/in/ashokit/entity/Product.java
package in.ashokit.entity;

import javax.persistence.Entity;
import javax.persistence.Id;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Entity
public class Product {

	@Id
	private Integer pid;
	private String pname;
	private Double price;
}