# swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "Api Documentation",
		"version": "1.0",
		"title": "Api Documentation",
		"termsOfService": "urn:tos",
		"contact": {
			
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0"
		}
	},
	"host": "localhost:1340",
	"basePath": "/",
	"tags": [
		{
			"name": "product-rest-controller",
			"description": "Product Rest Controller"
		}
	],
	"paths": {
		"/getProduct": {
			"get": {
				"tags": [
					"product-rest-controller"
				],
				"summary": "getProduct",
				"operationId": "getProductUsingGET",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/Product"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
		"/saveProduct": {
			"post": {
				"tags": [
					"product-rest-controller"
				],
				"summary": "saveProduct",
				"operationId": "saveProductUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "product",
						"description": "product",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Product"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		}
	},
	"definitions": {
		"Product": {
			"type": "object",
			"properties": {
				"pid": {
					"type": "integer",
					"format": "int32"
				},
				"pname": {
					"type": "string"
				},
				"price": {
					"type": "number",
					"format": "double"
				}
			},
			"title": "Product"
		}
	}
}